+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Compatibility with Node.JS and browsers by not using the `crypto.getRandomUUID` function
 
+### Fixed
+
+- Error message of `Target.broadcast`
+
 ## [0.1.2] - 2024-05-24
 
 ### Added