+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+- Allow setting fallback scripts (#1)
+
 ## [0.1.4] - 2024-07-05
 
 ### Added